splitted wsgi autoreload
the low lovel stuff is now in wsgi_autoreloading so it can easily
imported into a mod_wsgi loader script without invoking the
wsgi_autoreload top level code
